+++ src/app/plaintext/plaintext.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit, OnDestroy, ViewChild, ElementRef } from '@angular/core';
-import { RouterExtensions } from 'nativescript-angular/router';
+import { Component, OnInit, OnDestroy, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
+import { RouterExtensions, PageRoute } from 'nativescript-angular/router';
 import { registerElement } from 'nativescript-angular/element-registry';
 
 import { Subscription } from 'rxjs';
+import { switchMap, filter } from 'rxjs/operators';
 import { Page } from 'tns-core-modules/ui/page';
 import { Fab } from '@nstudio/nativescript-floatingactionbutton';
 import { PullToRefresh } from '@nstudio/nativescript-pulltorefresh';
@@ -21,7 +22,7 @@ registerElement('Fab', () => Fab);
     templateUrl: './plaintext.component.html',
     styleUrls: ['./plaintext.component.css'],
 })
-export class PlainTextComponent implements OnInit, OnDestroy {
+export class PlainTextComponent implements OnInit, OnDestroy, AfterViewInit {
 
     fileTitle: string;
     fileText: string;
@@ -31,8 +32,9 @@ export class PlainTextComponent implements OnInit, OnDestroy {
     fileTextView: ElementRef;
 
     constructor(
-        private page: Page,
         private routerExtensions: RouterExtensions,
+        private page: Page,
+        private pageRoute: PageRoute,
         private beancountFile: BeancountFileService,
         private sideDrawer: SideDrawerService,
     ) { }
@@ -52,6 +54,25 @@ export class PlainTextComponent implements OnInit, OnDestroy {
             // https://github.com/NativeScript/nativescript-angular/issues/1049
             this.ngOnDestroy();
         });
+    }
+
+    ngAfterViewInit() {
+        this.pageRoute.activatedRoute
+            .pipe(
+                switchMap(activatedRoute => activatedRoute.params),
+                filter(params => params.scroll === 'bottom'),
+            )
+            .forEach(() => {
+                // Wait until text view is loaded
+                // https://github.com/NativeScript/nativescript-angular/issues/1221#issuecomment-422813111
+                this.fileTextView.nativeElement.on('loaded', (event) => {
+                    // Use timeout to make it run on the next angular cycle
+                    // Otherwise scrollableHeight will be 0
+                    setTimeout(() => {
+                        this.scrollToBottom();
+                    }, 0);
+                });
+            });
     }
 
     onAddButtonLoaded(args) {